Store lowest arm num of record for record home page link
Report html view shows record ids with url for record home page.
Previously it always defaulted to the lowest numbered arm of the project
even if the record does not exist in that arm. It now keeps track of
which is the lowest arm for each record and uses this arm number in the
record home page url.